Please read on
- Moved first parts from OffchainStepper to HydratedRuntime to mimic contract
  design.
- Rename OffchainStepper to EVMRuntime
- Compute stackHash after execution step to combat OutOfMemory for large
  executions.
- Only compute changed elements of the stack.
- add primeTester unit test
- Merkelizer: use keccak directly in memHash & stackHash
  this gives us a huge perf gain and could be further improved
- some more things you can guess by viewing delta

Results:
runtime of PrimeTester before:
~7min
after:
29sec
