+++ README.md
@@ -53,11 +53,11 @@ Other contracts which makes use of the `EVMRuntime` are:
 It also exists a corresponding runtime implementation on the JS side.
 [You can take a look at the on-chain Verifier unit test on how it's used.](https://github.com/leapdao/solEVM-enforcer/blob/master/test/verifier.js)
 
-[The OffchainStepper mimics the EthereumRuntime contract](https://github.com/leapdao/solEVM-enforcer/blob/master/utils/OffchainStepper.js)
+[The off-chain EVMRuntime mimics the on-chain counterpart](https://github.com/leapdao/solEVM-enforcer/blob/master/utils/EVMRuntime.js)
 and together with the [Merkelizer](https://github.com/leapdao/solEVM-enforcer/blob/master/utils/Merkelizer.js),
 creates a Merkle Root of the individual execution steps (before and after each opcode) given `code`, `data` and other runtime properties.
 
-The `OffchainStepper` & `Merkelizer` are *WIP* too, as they get further improved and refactored to get closer to the `EVMRuntime` class model.
+The off-chain `EVMRuntime` & `Merkelizer` are *WIP* too, as they get further improved and refactored to get closer to the on-chain `EVMRuntime` class model.
 
 For the curious, there is also a [off-chain Dispute mock-implementation](https://github.com/leapdao/solEVM-enforcer/blob/master/utils/DisputeMock.js),
 with the same logic like the on-chain `Verifier` contract.
