+++ package.json
@@ -37,6 +37,7 @@
   "dependencies": {
     "ethereumjs-vm": "https://github.com/pinkiebell/ethereumjs-vm.git#release/v2.6.0-r6",
     "ethers": "=4.0.33",
+    "keccak": "=2.0.0",
     "openzeppelin-solidity": "=2.1.1"
   }
 }
