+++ test/fixtures/dispute.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const OffchainStepper = require('./../../utils/OffchainStepper');
+const HydratedRuntime = require('./../../utils/HydratedRuntime');
 const Merkelizer = require('./../../utils/Merkelizer');
 const OP = require('./../../utils/constants');
 const debug = require('debug')('dispute-test');
@@ -43,10 +43,10 @@ module.exports = (callback) => {
     let steps;
     let copy;
     let merkle;
-    const stepper = new OffchainStepper();
+    const runtime = new HydratedRuntime();
 
     beforeEach(async () => {
-      steps = await stepper.run({ code, data });
+      steps = await runtime.run({ code, data });
       copy = JSON.stringify(steps);
       merkle = new Merkelizer().run(steps, code, data);
     });
@@ -58,7 +58,7 @@ module.exports = (callback) => {
     it('challenger has an output error somewhere', async () => {
       const wrongExecution = JSON.parse(copy);
       wrongExecution[6].compactStack.push('0x0000000000000000000000000000000000000000000000000000000000000001');
-      wrongExecution[6].stack.push('0x0000000000000000000000000000000000000000000000000000000000000001');
+      wrongExecution[6].stackHash = '0x0000000000000000000000000000000000000000000000000000000000000001';
       const challengerMerkle = new Merkelizer().run(wrongExecution, code, data);
       await callback(code, data, merkle, challengerMerkle, 'solver');
     });
@@ -66,7 +66,7 @@ module.exports = (callback) => {
     it('solver has an output error somewhere', async () => {
       const wrongExecution = JSON.parse(copy);
       wrongExecution[6].compactStack.push('0x0000000000000000000000000000000000000000000000000000000000000001');
-      wrongExecution[6].stack.push('0x0000000000000000000000000000000000000000000000000000000000000001');
+      wrongExecution[6].stackHash = '0x0000000000000000000000000000000000000000000000000000000000000001';
       const solverMerkle = new Merkelizer().run(wrongExecution, code, data);
       await callback(code, data, solverMerkle, merkle, 'challenger');
     });
@@ -121,7 +121,7 @@ module.exports = (callback) => {
       const wrongExecution = JSON.parse(copy);
       for (let i = 1; i < wrongExecution.length; i += 2) {
         wrongExecution[i].compactStack.push('0x0000000000000000000000000000000000000000000000000000000000000000');
-        wrongExecution[i].stack.push('0x0000000000000000000000000000000000000000000000000000000000000000');
+        wrongExecution[i].stackHash = '0x0000000000000000000000000000000000000000000000000000000000000001';
       }
       const challengerMerkle = new Merkelizer().run(wrongExecution, code, data);
       await callback(code, data, merkle, challengerMerkle, 'solver');
@@ -131,7 +131,7 @@ module.exports = (callback) => {
       const wrongExecution = JSON.parse(copy);
       for (let i = 1; i < wrongExecution.length; i += 2) {
         wrongExecution[i].compactStack.push('0x0000000000000000000000000000000000000000000000000000000000000000');
-        wrongExecution[i].stack.push('0x0000000000000000000000000000000000000000000000000000000000000000');
+        wrongExecution[i].stackHash = '0x0000000000000000000000000000000000000000000000000000000000000001';
       }
       const solverMerkle = new Merkelizer().run(wrongExecution, code, data);
       await callback(code, data, solverMerkle, merkle, 'challenger');
@@ -199,10 +199,10 @@ module.exports = (callback) => {
     let steps;
     let copy;
     let merkle;
-    const stepper = new OffchainStepper();
+    const runtime = new HydratedRuntime();
 
     before(async () => {
-      steps = await stepper.run({ code });
+      steps = await runtime.run({ code });
       copy = JSON.stringify(steps);
       merkle = new Merkelizer().run(steps, code, data);
     });
@@ -255,12 +255,12 @@ module.exports = (callback) => {
 
     const data = '0x';
     let steps;
-    const stepper = new OffchainStepper();
+    const runtime = new HydratedRuntime();
     let challengerMerkle;
     let leaves;
 
     beforeEach(async () => {
-      steps = await stepper.run({ code });
+      steps = await runtime.run({ code });
       challengerMerkle = new Merkelizer().run(steps, code, data);
       leaves = JSON.stringify(challengerMerkle.tree[0]);
     });
@@ -375,10 +375,10 @@ module.exports = (callback) => {
     let steps;
     let copy;
     let merkle;
-    const stepper = new OffchainStepper();
+    const runtime = new HydratedRuntime();
 
     before(async () => {
-      steps = await stepper.run({ code, data });
+      steps = await runtime.run({ code, data });
       copy = JSON.stringify(steps);
       merkle = new Merkelizer().run(steps, code, data);
     });
@@ -450,10 +450,10 @@ module.exports = (callback) => {
 
     let steps;
     let merkle;
-    const stepper = new OffchainStepper();
+    const runtime = new HydratedRuntime();
 
     before(async () => {
-      steps = await stepper.run({ code, data });
+      steps = await runtime.run({ code, data });
       merkle = new Merkelizer().run(steps, code, data);
     });
 
@@ -479,10 +479,10 @@ module.exports = (callback) => {
     let steps;
     let merkle;
     let copy;
-    const stepper = new OffchainStepper();
+    const runtime = new HydratedRuntime();
 
     before(async () => {
-      steps = await stepper.run({ code, data });
+      steps = await runtime.run({ code, data });
       copy = JSON.stringify(steps);
       merkle = new Merkelizer().run(steps, code, data);
     });
@@ -513,10 +513,10 @@ module.exports = (callback) => {
     let steps;
     let merkle;
     let copy;
-    const stepper = new OffchainStepper();
+    const runtime = new HydratedRuntime();
 
     before(async () => {
-      steps = await stepper.run({ code, data });
+      steps = await runtime.run({ code, data });
       copy = JSON.stringify(steps);
       merkle = new Merkelizer().run(steps, code, data);
     });
