+++ utils/HydratedRuntime.js
@@ -1,0 +1,191 @@
+'use strict';
+
+const OP = require('./constants');
+const EVMRuntime = require('./EVMRuntime');
+const RangeProofHelper = require('./RangeProofHelper');
+const Merkelizer = require('./Merkelizer');
+
+const toHex = arr => arr.map(e => '0x' + e.toString(16).padStart(64, '0'));
+
+const OP_SWAP1 = parseInt(OP.SWAP1, 16);
+const OP_SWAP16 = parseInt(OP.SWAP16, 16);
+const OP_DUP1 = parseInt(OP.DUP1, 16);
+const OP_DUP16 = parseInt(OP.DUP16, 16);
+
+module.exports = class HydratedRuntime extends EVMRuntime {
+  async initRunState (obj) {
+    const runState = await super.initRunState(obj);
+    const stack = toHex(runState.stack);
+
+    runState.steps = [];
+    runState.prevStack = stack;
+    runState.stackHashes = Merkelizer.stackHashes(stack);
+
+    runState.memProof = new RangeProofHelper(runState.memory);
+    runState.callDataProof = new RangeProofHelper(runState.callData);
+    runState.codeProof = new RangeProofHelper(runState.code);
+    runState.callData = runState.callDataProof.proxy;
+    runState.memory = runState.memProof.proxy;
+    runState.code = runState.codeProof.proxy;
+
+    return runState;
+  }
+
+
+  async run (args) {
+    const runState = await super.run(args);
+
+    // a temporay hack for our unit tests :/
+    if (runState.steps.length > 0) {
+      runState.steps[runState.steps.length - 1].stack = toHex(runState.stack);
+    }
+
+    return runState.steps;
+  }
+
+  async runNextStep (runState) {
+    let pc = runState.programCounter;
+    const callDataProof = runState.callDataProof;
+    const codeProof = runState.codeProof;
+    const gasLeft = runState.gasLeft.addn(0);
+
+    callDataProof.reset();
+    codeProof.reset();
+
+    await super.runNextStep(runState);
+
+    const opcode = runState.opCode;
+    const returnData = '0x' + (runState.returnValue ? runState.returnValue.toString('hex') : '');
+
+    // if we have no errors and opcode is not RETURN or STOP, update pc
+    if (runState.errno === 0 && (opcode !== 0xf3 && opcode !== 0x00)) {
+      pc = runState.programCounter;
+    }
+
+    const codeReads = codeProof.readAndWrites.filter(
+      function (val) {
+        return val < runState.code.length;
+      }
+    ).sort(
+      function (a, b) {
+        if (a < b) {
+          return -1;
+        }
+        if (a > b) {
+          return 1;
+        }
+        return 0;
+      }
+    );
+
+    const step = {
+      callDataReadLow: callDataProof.readLow,
+      callDataReadHigh: callDataProof.readHigh,
+      codeReads: codeReads,
+      returnData: returnData,
+      pc: pc,
+      errno: runState.errno,
+      gasRemaining: runState.gasLeft.toNumber(),
+    };
+
+    this.calculateMemProof(runState, step);
+    this.calculateStackProof(runState, step);
+
+    runState.steps.push(step);
+  }
+
+  calculateMemProof (runState, step) {
+    const memProof = runState.memProof;
+    const prevMem = runState.prevMem;
+    const memSize = runState.memoryWordCount.toNumber();
+
+    // serialize the memory if it changed
+    if (memProof.readHigh !== -1 || memProof.writeHigh !== -1 || !prevMem || prevMem.length !== memSize) {
+      const mem = [];
+
+      const memStore = runState.memProof.data;
+      let i = 0;
+      while (i < memStore.length) {
+        const hexVal = Buffer.from(memStore.slice(i, i += 32)).toString('hex');
+        mem.push('0x' + hexVal.padEnd(64, '0'));
+      }
+      // fill the remaing zero slots
+      while (mem.length < memSize) {
+        mem.push(OP.ZERO_HASH);
+      }
+      step.mem = mem;
+      runState.prevMem = mem;
+    } else {
+      step.mem = prevMem;
+    }
+
+    step.memReadLow = memProof.readLow;
+    step.memReadHigh = memProof.readHigh;
+    step.memWriteLow = memProof.writeLow;
+    step.memWriteHigh = memProof.writeHigh;
+
+    memProof.reset();
+  }
+
+  calculateStackProof (runState, step) {
+    const opcode = runState.opCode;
+    let stackIn = runState.stackIn | 0;
+
+    if (opcode >= OP_SWAP1 && opcode <= OP_SWAP16) {
+      stackIn = (16 - (OP_SWAP16 - opcode)) * 2;
+    }
+
+    if (opcode >= OP_DUP1 && opcode <= OP_DUP16) {
+      stackIn = 16 - (OP_DUP16 - opcode);
+    }
+
+    // can happen on error - clip here
+    if (stackIn > runState.prevStack.length) {
+      stackIn = runState.prevStack.length;
+    }
+
+    // if stack changed
+    if (stackIn !== 0 || runState.prevStack.length !== runState.stack.length) {
+      // elements needed
+      step.compactStack = new Array(stackIn);
+
+      // remove the number of 'consumed' elements - if any
+      let old = stackIn + 0;
+      while (stackIn--) {
+        step.compactStack[stackIn] = runState.prevStack.pop();
+        runState.stackHashes.pop();
+      }
+
+      // add the new/changed elements - if any
+      const newElements = [];
+      for (let i = runState.prevStack.length; i < runState.stack.length; i++) {
+        let val = '0x' + runState.stack[i].toString(16).padStart(64, '0');
+        runState.prevStack.push(val);
+        newElements.push(val);
+      }
+      step.compactStackHash = runState.stackHashes[runState.stackHashes.length - 1];
+
+      const partialHashes = Merkelizer.stackHashes(newElements, step.compactStackHash);
+      // first element of partialHash is alread in the list
+      runState.stackHashes = runState.stackHashes.concat(partialHashes.slice(1, partialHashes.length));
+    } else {
+      step.compactStackHash = runState.stackHashes[runState.stackHashes.length - 1];
+      step.compactStack = [];
+    }
+
+    step.stackHash = runState.stackHashes[runState.stackHashes.length - 1];
+    step.stackSize = runState.stack.length;
+  }
+
+  async handleJUMP (runState) {
+    await super.handleJUMP(runState);
+
+    runState.codeProof.readAndWrites.push(runState.programCounter);
+  }
+
+  async handleJUMPI (runState) {
+    await super.handleJUMPI(runState);
+
+    runState.codeProof.readAndWrites.push(runState.programCounter);
+  }
+};
