+++ utils/Merkelizer.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const ethers = require('ethers');
+const createKeccakHash = require('keccak');
 
 const AbstractMerkleTree = require('./AbstractMerkleTree');
 const { ZERO_HASH } = require('./constants');
@@ -34,24 +35,50 @@ module.exports = class Merkelizer extends AbstractMerkleTree {
     return res;
   }
 
-  static stackHash (stack, sibling) {
-    let res = sibling || ZERO_HASH;
+  static stackHashes (stack, sibling) {
+    const res = sibling || ZERO_HASH;
+    const hashes = [res];
+    const len = stack.length;
 
-    for (var i = 0; i < stack.length; i++) {
-      res = ethers.utils.solidityKeccak256(
-        ['bytes32', 'bytes32'],
-        [res, stack[i]]
-      );
+    if (len === 0) {
+      return hashes;
     }
 
-    return res;
+    // could be further improved by a custom keccak implementation
+    const hash = createKeccakHash('keccak256');
+
+    hash.update(Buffer.from(res.replace('0x', ''), 'hex'));
+    for (let i = 0; i < len; i++) {
+      const val = Buffer.from(stack[i].replace('0x', ''), 'hex');
+
+      hash.update(val);
+
+      if (i !== len - 1) {
+        const buf = hash.digest();
+        hashes.push(`0x${buf.toString('hex')}`);
+        hash._finalized = false;
+        hash.update(buf);
+      }
+    }
+
+    hashes.push(`0x${hash.digest().toString('hex')}`);
+    return hashes;
+  }
+
+  static stackHash (stack, sibling) {
+    const res = this.stackHashes(stack, sibling);
+    return res[res.length - 1];
   }
 
   static memHash (mem) {
-    return ethers.utils.solidityKeccak256(
-      ['bytes32[]'],
-      [mem]
-    );
+    const len = mem.length;
+    const hash = createKeccakHash('keccak256');
+
+    for (let i = 0; i < len; i++) {
+      hash.update(Buffer.from(mem[i].replace('0x', ''), 'hex'));
+    }
+
+    return `0x${hash.digest().toString('hex')}`;
   }
 
   static dataHash (data) {
@@ -65,7 +92,7 @@ module.exports = class Merkelizer extends AbstractMerkleTree {
     // TODO: compact returnData
 
     if (!stackHash || stackHash === ZERO_HASH) {
-      stackHash = this.stackHash(execution.stack);
+      stackHash = this.stackHash(execution.stack || []);
     }
 
     if (!memHash || memHash === ZERO_HASH) {
@@ -112,11 +139,11 @@ module.exports = class Merkelizer extends AbstractMerkleTree {
 
     for (let i = 0; i < len; i++) {
       const exec = executions[i];
-      const stackHash = this.constructor.stackHash(exec.stack);
+      const stackHash = exec.stackHash;
 
       // convenience
-      exec.stackSize = exec.stack.length;
       exec.memSize = exec.mem.length;
+      exec.data = callData;
 
       // memory is changed if either written to or if it was expanded
       let memoryChanged = exec.memWriteLow !== -1;
