+++ utils/index.js
@@ -2,7 +2,8 @@
 
 const Constants = require('./constants.js');
 const AbstractMerkleTree = require('./AbstractMerkleTree.js');
-const OffchainStepper = require('./OffchainStepper.js');
+const EVMRuntime = require('./EVMRuntime.js');
+const HydratedRuntime = require('./HydratedRuntime.js');
 const Merkelizer = require('./Merkelizer.js');
 const ProofHelper = require('./ProofHelper.js');
 const ExecutionPoker = require('./ExecutionPoker.js');
@@ -11,7 +12,8 @@ const FragmentTree = require('./FragmentTree.js');
 module.exports = {
   Constants,
   AbstractMerkleTree,
-  OffchainStepper,
+  EVMRuntime,
+  HydratedRuntime,
   Merkelizer,
   ProofHelper,
   ExecutionPoker,
